backend/requirements.txt
fastapi
uvicorn
paddlepaddle
paddleocr
sqlalchemy
psycopg2-binary
python-multipart

